requirements.txt
nltk
bs4
requests
ipykernel
python-box==6.0.2
pyYAML
ensure==1.0.2
types-pyYAML
pandas
openpyxl